# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cauteum Gateway REST API
  version: 0.1.0
  summary: Client-facing contract for the gateway control-plane REST API.
  description: >-
    This specification documents the client-facing REST control-plane surface:
    health, gateway metadata, sandbox registry, logs, policy proposals,
    providers, workspaces, services, templates, settings, inference, and local
    or OIDC authentication configuration. Operational probes, diagnostics,
    streaming relay transports, compatibility aliases, and RPC are excluded.
    Sandbox registry writes do not provision or stop a compute runtime; runtime
    lifecycle operations are defined by the OpenShell RPC contract.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:7443
    description: Local development gateway. Configure HTTPS for remote use.
...
security:
  - bearerAuth: []
paths:
...
  /v1/providers/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getProvider
      tags: [Providers]
      summary: Read provider metadata
      description: Credential values are never returned. Requires provider:read when OIDC scope authorization is enabled.
      responses:
        '200':
          description: Provider metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: upsertProvider
      tags: [Providers]
      summary: Create or replace a provider instance
      description: Credential values are write-only and stored encrypted when local credential storage is configured. Requires provider:write when OIDC scope authorization is enabled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderWrite'
      responses:
        '204':
          description: Provider was saved.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: Credential encryption service is unavailable.
          content:
            text/plain:
              schema:
                type: string
    delete:
      operationId: deleteProvider
      tags: [Providers]
      summary: Delete a provider instance
      description: Requires provider:write when OIDC scope authorization is enabled.
      responses:
        '204':
          description: Provider was deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/inference:
    get:
      operationId: getInferenceRoute
      tags: [Configuration]
      summary: Read the configured inference route
      responses:
        '200':
          description: Inference route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InferenceRoute'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: setInferenceRoute
      tags: [Configuration]
      summary: Configure the inference route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRouteInput'
      responses:
        '200':
          description: Saved inference route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InferenceRoute'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      operationId: clearInferenceRoute
      tags: [Configuration]
      summary: Clear the inference route
      responses:
        '204':
          description: Inference route was cleared.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/settings:
    get:
      operationId: listSettings
      tags: [Configuration]
      summary: List gateway settings
      responses:
        '200':
          description: Settings keyed by name.
          content:
            application/json:
              schema:
                type: object
                required: [settings]
                properties:
                  settings:
                    type: object
                    additionalProperties:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/settings/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getSetting
      tags: [Configuration]
      summary: Read one gateway setting
      responses:
        '200':
          description: Setting value.
          content:
            application/json:
              schema:
                type: object
                required: [key, value]
                properties:
                  key:
                    type: string
                  value:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: setSetting
      tags: [Configuration]
      summary: Set one gateway setting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: string
      responses:
        '204':
          description: Setting was saved.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      operationId: deleteSetting
      tags: [Configuration]
      summary: Delete one gateway setting
      responses:
        '204':
          description: Setting was deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/templates:
    get:
      operationId: listTemplates
      tags: [Configuration]
      summary: List workload templates
      responses:
        '200':
          description: Template records.
          content:
            application/json:
              schema:
                type: object
                required: [templates]
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/templates/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getTemplate
      tags: [Configuration]
      summary: Read a workload template
      responses:
        '200':
          description: Template record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: upsertTemplate
      tags: [Configuration]
      summary: Create or replace a workload template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Template'
      responses:
        '204':
          description: Template was saved.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteTemplate
      tags: [Configuration]
      summary: Delete a workload template
      responses:
        '204':
          description: Template was deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/services:
    get:
      operationId: listServices
      tags: [Services]
      summary: List registered services
      responses:
        '200':
          description: Service records.
          content:
            application/json:
              schema:
                type: object
                required: [services]
                properties:
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/services/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getService
      tags: [Services]
      summary: Read a registered service
      responses:
        '200':
          description: Service record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: upsertService
      tags: [Services]
      summary: Create or replace a registered service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInput'
      responses:
        '200':
          description: Saved service record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteService
      tags: [Services]
      summary: Delete a registered service
      responses:
        '204':
          description: Service was deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workspaces:
    get:
      operationId: listWorkspaces
      tags: [Workspaces]
      summary: List workspaces
      responses:
        '200':
          description: Workspace records.
          content:
            application/json:
              schema:
                type: object
                required: [workspaces]
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createWorkspace
      tags: [Workspaces]
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Created workspace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workspaces/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getWorkspace
      tags: [Workspaces]
      summary: Read a workspace
      responses:
        '200':
          description: Workspace record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteWorkspace
      tags: [Workspaces]
      summary: Delete an empty workspace
      responses:
        '204':
          description: Workspace was deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Workspace still contains resources.
          content:
            text/plain:
              schema:
                type: string
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/whoami:
    get:
      operationId: getCurrentIdentity
      tags: [Gateway]
      summary: Read the authenticated caller identity
      responses:
        '200':
          description: Caller identity and authorization context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/auth/login:
    get:
      operationId: startLocalDevelopmentLogin
      tags: [Gateway]
      summary: Obtain a local development gateway token
      description: >-
        Available only to loopback clients. If redirect_uri is supplied, the
        token is placed in its query string; this is intended only for a local
        loopback callback and should not be copied into logs or browser history.
      security: []
      parameters:
        - name: redirect_uri
          in: query
          required: false
          schema:
            type: string
            format: uri
          description: Optional http://127.0.0.1 callback URL.
      responses:
        '200':
          description: Local development token response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalLogin'
        '302':
          description: Redirect to the supplied loopback URL with token query parameter.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or local gateway token
...
  schemas:
...
    Provider:
      type: object
      description: Provider metadata. Credential values are not included in responses.
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        workspace:
          type: string
        env_vars:
          type: array
          items:
            type: string
        credential_driver:
          type: string
        credential_handles:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CredentialHandle'
        credential_expires_at_ms:
          type: object
          additionalProperties:
            type: integer
            format: int64
        runtime_credentials:
          type: boolean
        config:
          type: object
          additionalProperties:
            type: string
        refresh:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProviderRefreshConfig'
    ProviderWrite:
      allOf:
        - $ref: '#/components/schemas/Provider'
        - type: object
          properties:
            credentials:
              type: object
              writeOnly: true
              additionalProperties:
                type: string
    CredentialHandle:
      type: object
      required: [driver, handle]
      properties:
        driver:
          type: string
        handle:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
...
    InferenceRoute:
      type: object
      required: [provider]
      properties:
        provider:
          type: string
        model:
          type: string
        timeout_sec:
          type: integer
        version:
          type: integer
    InferenceRouteInput:
      type: object
      required: [provider, model]
      properties:
        provider:
          type: string
        model:
          type: string
        timeout_sec:
          type: integer
...
    Service:
      type: object
      required: [name, sandbox, port, backend_host, backend_port, updated_at]
      properties:
        name:
          type: string
        sandbox:
          type: string
        port:
          type: integer
        backend_host:
          type: string
        backend_port:
          type: integer
        updated_at:
          type: string
          format: date-time
    ServiceInput:
      type: object
      properties:
        name:
          type: string
        sandbox:
          type: string
        port:
          type: integer
        backend_host:
          type: string
        backend_port:
          type: integer
...
  responses:
    BadRequest:
      description: Invalid request. Current response body is text/plain.
      content:
        text/plain:
          schema:
            type: string
    Unauthorized:
      description: Missing or invalid bearer token. Current response body is text/plain.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        text/plain:
          schema:
            type: string
    Forbidden:
      description: Caller is not authorized. Current response body is text/plain.
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource not found. Current response body is text/plain.
      content:
        text/plain:
          schema:
            type: string
    InternalError:
      description: Gateway error. Current response body is text/plain.
      content:
        text/plain:
          schema:
            type: string